@@ -5,7 +5,7 @@ resolver = "2"
 [workspace.package]
 version = "0.1.0"
 edition = "2024"
-rust-version = "1.86.0"
+rust-version = "1.88.0"
 authors = ["bug-ops"]
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/bug-ops/feedparser-rs"
