@@ -1,15 +1,6 @@
-/// Compatibility utilities for feedparser API
-///
-/// This module provides utilities to ensure API compatibility with
-/// Python's feedparser library.
+// Compatibility utilities for feedparser API
+//
+// This module provides utilities to ensure API compatibility with
+// Python's feedparser library.
 
 // TODO: Implement in later phases as needed
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn test_placeholder() {
-        // Placeholder test
-        assert!(true);
-    }
-}
