@@ -1,15 +1,6 @@
-/// Text processing utilities
-///
-/// This module will provide functions for text manipulation,
-/// such as trimming, normalizing whitespace, etc.
+// Text processing utilities
+//
+// This module will provide functions for text manipulation,
+// such as trimming, normalizing whitespace, etc.
 
 // TODO: Implement as needed
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn test_placeholder() {
-        // Placeholder test
-        assert!(true);
-    }
-}
